Refactored some code
