+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "boilerplate-frontend-react-graphql-apollo",
-  "version": "0.4.48",
+  "version": "0.4.49",
   "description": "boilerplate-frontend-react-graphql-apollo",
   "private": true,
   "dependencies": {
