+++ src/AuthContext.js
@@ -2,7 +2,7 @@ import React, { createContext, useState } from 'react'
 
 import jwt from 'jsonwebtoken'
 
-import { saveSession, recoverSession, deleteSession, storeUserDataOnSessionStorage, recoverUserDataFromSessionStorage, deleteUserDataFromSessionStorage } from './utils/utils'
+import { saveSession, recoverSession, deleteSession, storeUserDataOnSessionStorage, recoverUserDataFromSessionStorage, deleteUserDataFromSessionStorage } from './utils/session'
 
 export const AuthContext = createContext()
